// validation/models/WarningCategory.ts
export enum WarningCategory {
  INFERRED_LATCH = 'INFERRED_LATCH',
  UNUSED_SIGNAL = 'UNUSED_SIGNAL',
  MULTI_DRIVEN_NET = 'MULTI_DRIVEN_NET',
  RESOURCE_HIGH_UTILIZATION = 'RESOURCE_HIGH_UTILIZATION',
  UNSUPPORTED_SYNTHESIS_PRAGMA = 'UNSUPPORTED_SYNTHESIS_PRAGMA',
  BITWIDTH_TRUNCATION = 'BITWIDTH_TRUNCATION',
  CLOCK_DOMAIN_CROSSING = 'CLOCK_DOMAIN_CROSSING',
  COMBINATIONAL_LOOP = 'COMBINATIONAL_LOOP',
  UNINITIALIZED_REGISTER = 'UNINITIALIZED_REGISTER',
  DEPRECATED_SYNTAX = 'DEPRECATED_SYNTAX',
  HIGH_FANOUT = 'HIGH_FANOUT'
}

export interface WarningDetail {
  category: WarningCategory;
  message: string;
  file?: string;
  line?: number;
  column?: number;
  tool: string;
  suggestion?: string;
  rawOutput?: string;
}